JavaApplication3/src/javaapplication3/OrderInfo.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package javaapplication3;

/**
 *
 * @author Aremu Oluwaseyi Festus(https://github.com/Dev-Rem)
 */
public class OrderInfo {
    private String comment;
    private String orderType;

    public OrderInfo(){}
    
    public OrderInfo(String comment, String orderType) {
        this.comment = comment;
        this.orderType = orderType;
    }

    public String getComment() {
        return comment;
    }

    public void setComment(String comment) {
        this.comment = comment;
    }

    public String getOrderType() {
        return orderType;
    }

    public void setOrderType(String orderType) {
        this.orderType = orderType;
    }

    
}
